BLD: pygeos isn't an optional dependency anymore (#694)
* Turn pygeos to second-optional dependency

* update example

* correct variable name

* update example
