+++ environment-dev.yaml
@@ -9,7 +9,7 @@ dependencies:
   # optional
   - scikit-learn >= 1.1
   - geopandas >= 0.10.0
-  - pygeos >= 0.11.1
+  - pygeos
   - geopy
   - jenkspy
   - topojson
