+++ setup.cfg
@@ -48,8 +48,6 @@ python_requires = >=3.8
 opt =
     scikit-learn >= 1.1; python_version >= "3.8"
     geopandas >= 0.10.0; python_version >= "3.8"
-    pygeos >= 0.8; python_version <= "3.9"
-    pygeos >= 0.11.1; python_version >= "3.10"
 test =
     pytest
     pytest-xdist
